requirements.txt
langgraph
langchain_openai
deepeval
python-dotenv
openai